- installed workaround for bug 6654 with log for observation over time      (http://bugzilla.osafoundation.org/show_bug.cgi?id=6654)      (restore settings feature  only restores 1 of 2 calendars)
git-svn-id: http://svn.osafoundation.org/chandler/trunk@11690 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/repository/item/Item.py
@@ -1436,9 +1436,10 @@ class Item(CItem):
         """
 
         if cloudAlias is not None:
-            clouds = self._kind.getClouds(cloudAlias)
-            for cloud in clouds:
-                cloud.deleteItems(self, recursive, cloudAlias)
+            if not self.isDeleted():
+                clouds = self._kind.getClouds(cloudAlias)
+                for cloud in clouds:
+                    cloud.deleteItems(self, recursive, cloudAlias)
 
         elif not self._status & (Item.DELETED | Item.DELETING | Item.DEFERRING):
